support for long literals in native calls

+++ Java-Core/src/main/java/prompto/csharp/CSharpIntegerLiteral.java
@@ -4,11 +4,11 @@ package prompto.csharp;
 
 public class CSharpIntegerLiteral extends CSharpLiteral {
 
-	Integer value;
+	Long value;
 	
 	public CSharpIntegerLiteral(String text) {
 		super(text);
-		this.value = Integer.valueOf(text);
+		this.value = Long.valueOf(text);
 	}
 
 	@Override
