+++ Java-Core/src/main/java/prompto/java/JavaIntegerLiteral.java
@@ -7,21 +7,21 @@ import prompto.type.IType;
 
 public class JavaIntegerLiteral extends JavaLiteral {
 
-	Integer value;
+	Long value;
 	
 	public JavaIntegerLiteral(String text) {
 		super(text);
-		this.value = Integer.valueOf(text);
+		this.value = Long.valueOf(text);
 	}
 
 	@Override
-	public Integer interpret(Context context) {
+	public Long interpret(Context context) {
 		return value;
 	}
 	
 	@Override
 	public IType check(Context context) throws SyntaxError {
-		return new JavaClassType(Integer.class);
+		return new JavaClassType(Long.class);
 	}
 	
 	@Override
